chore: update imports for scheduler again
